Seperated tutorial from setup for cmf_lumped and added a GUI example
